src/requirements.txt
aiomysql==0.2.0
aiosqlite==0.18.0
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.95.2
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.46
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
